src/test/java/facades/MovieFacadeTest.java
package facades;

import utils.EMF_Creator;
import entities.*;
import javax.persistence.EntityManager;
import javax.persistence.EntityManagerFactory;
import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.AfterEach;
import static org.junit.jupiter.api.Assertions.assertEquals;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Test;
import utils.Settings;
import utils.EMF_Creator.DbSelector;
import utils.EMF_Creator.Strategy;

//Uncomment the line below, to temporarily disable this test
//@Disabled
public class MovieFacadeTest {

    private static EntityManagerFactory emf;
    private static MovieFacade facade;

    public MovieFacadeTest() {
    }

    //@BeforeAll
    public static void setUpClass() {
        emf = EMF_Creator.createEntityManagerFactory(
                "pu",
                "jdbc:mysql://localhost:3307/startcode_test",
                "dev",
                "ax2",
                EMF_Creator.Strategy.CREATE);
        facade = MovieFacade.getFacadeExample(emf);
    }

    /*   **** HINT **** 
        A better way to handle configuration values, compared to the UNUSED example above, is to store those values
        ONE COMMON place accessible from anywhere.
        The file config.properties and the corresponding helper class utils.Settings is added just to do that. 
        See below for how to use these files. This is our RECOMENDED strategy
     */
    @BeforeAll
    public static void setUpClassV2() {
       emf = EMF_Creator.createEntityManagerFactory(DbSelector.TEST,Strategy.DROP_AND_CREATE);
       facade = MovieFacade.getFacadeExample(emf);
    }

    @AfterAll
    public static void tearDownClass() {
//        Clean up database after test is done or use a persistence unit with drop-and-create to start up clean on every test
    }

    // Setup the DataBase in a known state BEFORE EACH TEST
    //TODO -- Make sure to change the script below to use YOUR OWN entity class
       private final  Movie l =new Movie(1999, "In the Deep", new String[]{"Cameron","Karry","Davis"});
    @BeforeEach
    public void setUp() {
        EntityManager em = emf.createEntityManager();
        try {
            em.getTransaction().begin();
            em.createNamedQuery("Movie.deleteAllRows").executeUpdate();
            em.persist(l);
            em.persist(new Movie(1998, "Down under", new String[]{"Kaden","Barry","Lero"}));
            em.getTransaction().commit();
        } finally {
            em.close();
      }
    }
    

    @AfterEach
    public void tearDown() {
//        Remove any data after each test was run
    }

    // TODO: Delete or change this method 
    @Test
    public void testCountFacadeMethod() {
        assertEquals(2, facade.count(), "Expects two rows in the database");
    }
    @Test
    public void testGetByNameFacadeMethod() {
        Movie expected = new Movie(1998, "Down under", new String[]{"Kaden","Barry","Lero"});
        assertEquals(expected.getActors()[0], facade.findByTitle("Down under").get(0).getActors()[0]);
    }
    @Test
    public void testGetByIdFacadeMethod() {
        Movie expected = new Movie(1999, "In the Deep", new String[]{"Cameron","Karry","Davis"});
        assertEquals(expected.getName(), facade.findByID(l.getId()).getName());
    }
}
